requirements.txt
asgiref==3.7.2
boto3==1.20.26
botocore==1.23.54
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==41.0.7
decorator==4.4.2
dj-database-url==2.1.0
Django==4.2.7
django-anymail==10.2
django-ckeditor-5==0.2.13
django-cors-headers==3.14.0
django-jazzmin==2.6.0
django-storages==1.12.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
environs==10.0.0
gunicorn==21.2.0
idna==3.6
imageio==2.34.1
imageio-ffmpeg==0.5.1
inflection==0.5.1
jmespath==0.10.0
marshmallow==3.20.1
moviepy==1.0.3
numpy==2.0.0
packaging==23.2
pillow==10.3.0
proglog==0.1.10
psycopg2==2.9.9
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.5.2
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.4
stripe==7.9.0
tqdm==4.66.4
typing_extensions==4.9.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.18
